plotting of global timeseries works
